device-tree: Enhancement: "compatible = " is now a list
To avoid having to update the linux kernel every time a Xilinx core changes
it's version string, with this change, the generator now generates
"compatible=" lines with:
        1) the core's actual version
        2) the oldest core version that is software compatible with the core's
           actual version.
Example:
        compatible = "xlnx,xps-intc-2.00.a", "xlnx,xps-intc-1.00.a";

There are a few rules that go with this:
1) The oldest core version listed must be a version that is already supported
   by the linux kernel, or this change doesn't help.
2) The versions are considered compatible if the driver/kernel code made for
   the older version will work with the core's actual version. What this means
   is that the newer core might have additional features, but the core must
   be operable by the driver/kernel using only the older feature set.

This feature is implemented by using a static table in the tcl script. At this
time there does not seem to be a good  method for determining this using
automation.

Signed-off-by: John Bonesio <john.bonesio@xilinx.com>