device-tree: Support first override parameter for gpio-reset IP
MSS fragment
 PARAMETER periph_type_overrides = {hard-reset-gpios Reset_GPIO 1 1}

DTS fragment
	hard-reset-gpios = < &Reset_GPIO 1 1 >;

Implementation description:
I created new type called "labelref-ext" which is able to generate
dts line.
  parameter = < &name 1st 2nd 3rd .. >;

BSP check if that gpio IP name is valid for this design and if gpio
pin not exceed gpio width. Pin started from 1 to GPIO width.
For this reset usage BSP doesn't take care about last parameter which
is size of gpio pins - it is hardcoded to always setup size to 1.

Signed-off-by: Michal Simek <monstr@monstr.eu>


git-svn-id: https://developer.petalogix.com/pr/internal/petalinux/branches/petalinux-v1.00-devel@6981 252c5255-be06-0410-b2d7-f4a01eabda39