Xilinx: DT Generator: Support AXI interface on new Microblaze
Update the DT Generator to handle the new AXI interface on
Microblaze v8.00 and also the AXI IP connected to this i/f.
This patch also splits all the system info which goes into
the dts, into 2 parts (1. AXI IP info, 2. rest of the system
info), such that the AXI IP can be placed at the end of the
dts file and make it look similar to the other dts files (the
write() function sorts the info into alphabetical order before
writing it into the dts file and if we don't split it into 2
parts, we'll end up with 'A'XI IP at the top of the file)

Signed-off-by: Sadanand M <sadanan@xilinx.com>