BSP: device-tree: Fix bsp to support systems with 2GB ram.
Size is int ( not unsigned int) that's why was necessary to
convert it to hex and compare hex.
There were also problem because MPMC CTRL code (for ll_temac)
depends on it (maybe accidentaly). I added detection for
baseaddr eguals 0xffffffff which is default value for MPMC_CTRL_BASEADDR
if CTRL interface is not defined.


git-svn-id: https://developer.petalogix.com/pr/internal/petalinux/branches/petalinux-v1.00-devel@8622 252c5255-be06-0410-b2d7-f4a01eabda39