Xilinx: DT Generator: Add MDIO support for ethernetlite
Create a phy node for ethernetlite devices, if MDIO bus is
present. This patch uses a phy-addr of 7 (default on most
Xilinx boards) for all the phy-nodes. The user has to edit
the dts file if the phy addr is different. Currently, there
is no way for this tcl to find the phy addr automatically

Signed-off-by: Sadanand M <sadanan@xilinx.com>