Generate proper platform name - parsing petalinux_bsp_? string.
If the project has that name then it is SDK and look for
name for hw design name.

I think we shoudl do it in a little bit smarter way.

On the other hand use which create hw platform by name
petalinux_bsp_? is ...

Systems with more than 10 cpus are not supported.


git-svn-id: https://developer.petalogix.com/pr/internal/petalinux/branches/petalinux-v1.00-devel@8907 252c5255-be06-0410-b2d7-f4a01eabda39