dts-generator: tag UART-capable MDMs as serial
New logic in petalinux-qemu-boot for auto serial mapping requires serial
devices to be in the DTS as serial@<addr>

This is a quick hack to export any MDMs that have UART functionality to be
serial@ devices, thus they work in the new unified serial framework.

There is no impact here because there are no kernel drivers looking for debug@
nodes anyway.

Signed-off-by: John Williams <john.williams@petalogix.com>