DTS/uboot TCL:allow to use axi_emc as main memory
Another TCL script petalinux_v2_00_a/data/petalinux_v2_1_0.mld is also required
to add axi_emc to the main_memory list. But once I added axi_emc to the list
it is possible the XSDK will auto select the Flash as the main memory if the
user use XSDK to create fs-boot Xilinx C project. This will requires board
bring up document modification. And the user will notice this difference when
they use new version of PetaLinux, since if it is the Flash selected as the
main memory, the BSP generation will error out. The users will need to
manually make sure it is the memory controller selected as the main memory.

Signed-off-by: Wendy Liang <wendy.liang@petalogix.com>