device-tree tcl: Add support for Xylon IPs
We use Xylon IPs for our demo and it is just easier
to be able to generate device-tree for it.

There is one really bad idea which is that register
range is not C_AXI_ or C_ BASE/HIGHADDR but they use
specific symbols C_REGS_ prefix instead.

There are also other parameters which names C_VMEM_ and C_BB_
for specific purpose.

Default case can't handle these IPs because they are
not written in standard way.

Signed-off-by: Michal Simek <michal.simek@petalogix.com>