dt tcl: AXI stream: generated phandles by masters
Generated phandler for masters of axi-stream rather than slaves. This is more
consistent with the way phandles are user for interrupts, I.E. A generator of
an interrupt has a phandle to an intc so a generator (master) of an axi-stream
transaction has a phandle to its sink (slave).

No net effect for the usual Ethernet<->DMA setup so no kernel breakage, but
future proofs DTG for use with QEMU axi-stream for other devices.

Signed-off-by: Peter A. G. Crosthwaite <peter.crosthwaite@petalogix.com>