device-tree-generator: stdout peripherals are now selectable in configuration.
During the configuration for device tree generation, the user was required
to manually type the peripheral name in the stdout field. This was prone
to typos and errors.

Instead, the generator now detects stdout compatible peripherals and the user
can now select one of them from a drop-down menu in the stdout field during
configuration.