smcc: Generate Nand/Nor/Sram as subnode of smcc
Generate smcc as the last ip in the list which ensure
that we can store all nand/nor/sram node descriptions
which we can add in smcc node.

Also replace ps7_sram by ps7_nor if we are working with flash.
It is not the best option to show ps7_sram if we know that it is nor flash.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>