Fix the problem when IP has BASE/HIGH addresses pairs
For IPs which has not specific handler change logic
for looking for addresses. Directly looking for
address pairs and setup ranges list.
In case that IP has a lot of BASE/HIGH pairs but
only one is used.
Previous implementation detect number of pairs
but didn't check if all of them are valid.
Current implementation only working with valid
configurations.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>