ps7_sram: Remove the incorrect assumption on ps7_sram node
ps7_sram_* are refer to NOR flash in 14.4. Removes the if statement on
the assumption of ps_sram_0 is nor and ps7_sram_1 is sram. Note, this is
also the wrong assumption in 14.4.

Rename the node name to primary_flash if it is set as primary flash in
the menu options.

Signed-off-by: Jason Wu <huanyu@xilinx.com>