Generate ps7_sdio clock-frequency again
Here is confusion in device-tree which is that sometimes
clock-frequency is output freq instead of input freq.
For sdhci is this property required by shdci-pltfm.c:
sdhci_get_of_property and there is no way how to
pass this clock frequency directly from the driver
because structure is not initialized in driver code.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>