Add enet/DMA control-stream
The AXI Ethernet and DMA have not one but two AXI Stream connections between
them. To make the dtb more accurate both these linkages should be in the dts
independently.

This initial patch is about getting the syntax right in the normal use case,
where both streams run in parallel from a single Ethernet to a single DMA.
But what we really need to do long term is check the control and data stream
independently and potentially even connect them to different devices.

My motivation here is QEMU. QEMU now has support for multiple stream connections
to the one device and the FDT driven machine model instantiation wants this.
Othewise we need specific "if AXI DMA ... " logic to tell QEMU to interpret the
axistream-connected phandle linkage as not one but two streams.

Should be zero kernel side impact as is a purely additive dts change.

Signed-off-by: Peter Crosthwaite <peter.crosthwaite@xilinx.com>
Acked-by: Michal Simek <michal.simek@xilinx.com>