Do not generate several memory nodes in root
When system contains more than one memory controller
all of them were generated to DTS root which is incorrect.
Find out the first controller based on this order
in get_first_mem_controller().

PetaLinux flow is not affected because memory node
is chosen in high level BSP and pass to device-tree BSP.
This patch change behavior also when memory node is not selected.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Acked-by: Peter Crosthwaite <peter.crosthwaite@xilinx.com>