Create workaround for axi_dma node generation
If axi_ethernet is using axi_dma axistream-connected
and axistream-control-connected must be generated.
If axistream is not connected to axi_ethernet IP
then different device node is generated.

This is workaround till axi_ethernet driver is fixed
to use dma framework.

axis_loopback_widget are used for setting up loopback mode
but this IP is not generated.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Acked-by: Peter Crosthwaite <peter.crosthwaite@xilinx.com>